+++ mayan/apps/document_indexing/queues.py
@@ -1,11 +1,13 @@
 from django.utils.translation import ugettext_lazy as _
 
-from mayan.apps.common.queues import queue_tools
 from mayan.apps.task_manager.classes import CeleryQueue
-from mayan.apps.task_manager.workers import worker_b
+from mayan.apps.task_manager.workers import worker_b, worker_c
 
 queue_indexing = CeleryQueue(
     label=_('Indexing'), name='indexing', worker=worker_b
+)
+queue_indexing_slow = CeleryQueue(
+    label=_('Indexing slow'), name='indexing_slow', worker=worker_c
 )
 
 queue_indexing.add_task_type(
@@ -16,7 +18,8 @@ queue_indexing.add_task_type(
     label=_('Index document'),
     dotted_path='mayan.apps.document_indexing.tasks.task_index_instance_document_add'
 )
-queue_tools.add_task_type(
+
+queue_indexing_slow.add_task_type(
     label=_('Rebuild index'),
     dotted_path='mayan.apps.document_indexing.tasks.task_index_template_rebuild'
 )