+++ mayan/apps/document_signatures/queues.py
@@ -1,11 +1,13 @@
 from django.utils.translation import ugettext_lazy as _
 
-from mayan.apps.common.queues import queue_tools
 from mayan.apps.task_manager.classes import CeleryQueue
-from mayan.apps.task_manager.workers import worker_c
+from mayan.apps.task_manager.workers import worker_c, worker_d
 
 queue_signatures = CeleryQueue(
     label=_('Signatures'), name='signatures', worker=worker_c
+)
+queue_signatures_slow = CeleryQueue(
+    label=_('Signatures slow'), name='signatures_slow', worker=worker_d
 )
 
 queue_signatures.add_task_type(
@@ -21,11 +23,11 @@ queue_signatures.add_task_type(
     label=_('Verify document file')
 )
 
-queue_tools.add_task_type(
+queue_signatures_slow.add_task_type(
     dotted_path='mayan.apps.document_signatures.tasks.task_verify_missing_embedded_signature',
     label=_('Verify missing embedded signature')
 )
-queue_tools.add_task_type(
+queue_signatures_slow.add_task_type(
     dotted_path='mayan.apps.document_signatures.tasks.task_refresh_signature_information',
     label=_('Refresh existing signature information')
 )