+++ mayan/apps/duplicates/queues.py
@@ -1,12 +1,13 @@
 from django.utils.translation import ugettext_lazy as _
 
-from mayan.apps.common.queues import queue_tools
 from mayan.apps.task_manager.classes import CeleryQueue
-from mayan.apps.task_manager.workers import worker_b
-
+from mayan.apps.task_manager.workers import worker_b, worker_c
 
 queue_duplicates = CeleryQueue(
     label=_('Duplicates'), name='duplicates', worker=worker_b
+)
+queue_duplicates_slow = CeleryQueue(
+    label=_('Duplicates slow'), name='duplicates_slow', worker=worker_c
 )
 
 queue_duplicates.add_task_type(
@@ -18,7 +19,7 @@ queue_duplicates.add_task_type(
     label=_('Scan document duplicates')
 )
 
-queue_tools.add_task_type(
+queue_duplicates_slow.add_task_type(
     dotted_path='mayan.apps.duplicates.tasks.task_duplicates_scan_all',
     label=_('Duplicated document scan')
 )