+++ mayan/apps/file_caching/queues.py
@@ -1,11 +1,13 @@
 from django.utils.translation import ugettext_lazy as _
 
-from mayan.apps.common.queues import queue_tools
 from mayan.apps.task_manager.classes import CeleryQueue
-from mayan.apps.task_manager.workers import worker_b
+from mayan.apps.task_manager.workers import worker_b, worker_c
 
 queue_file_caching = CeleryQueue(
     name='file_caching', label=_('File caching'), worker=worker_b
+)
+queue_file_caching_slow = CeleryQueue(
+    name='file_caching_slow', label=_('File caching slow'), worker=worker_c
 )
 
 queue_file_caching.add_task_type(
@@ -13,7 +15,7 @@ queue_file_caching.add_task_type(
     label=_('Purge a file cache partition')
 )
 
-queue_tools.add_task_type(
+queue_file_caching_slow.add_task_type(
     dotted_path='mayan.apps.file_caching.tasks.task_cache_purge',
     label=_('Purge a file cache')
 )