+++ mayan/apps/storage/queues.py
@@ -3,15 +3,23 @@ from datetime import timedelta
 from django.utils.translation import ugettext_lazy as _
 
 from mayan.apps.task_manager.classes import CeleryQueue
-from mayan.apps.task_manager.workers import worker_d
+from mayan.apps.task_manager.workers import worker_c, worker_d
 
 from .literals import (
     TASK_DOWNLOAD_FILE_STALE_INTERVAL, TASK_SHARED_UPLOADS_STALE_INTERVAL
 )
 
+queue_storage = CeleryQueue(
+    label=_('Storage'), name='storage', worker=worker_c
+)
 queue_storage_periodic = CeleryQueue(
     label=_('Storage periodic'), name='storage_periodic', transient=True,
     worker=worker_d
+)
+
+queue_storage.add_task_type(
+    dotted_path='mayan.apps.storage.tasks.task_shared_upload_delete',
+    label=_('Delete a shared upload'), name='task_shared_upload_delete'
 )
 
 queue_storage_periodic.add_task_type(