+++ mayan/apps/storage/tasks.py
@@ -35,3 +35,23 @@ def task_shared_upload_stale_delete():
     SharedUploadedFile.objects.stale_delete()
 
     logger.debug('Finished')
+
+
+@app.task(bind=True, ignore_result=True, retry_backoff=True)
+def task_shared_upload_delete(self, shared_uploaded_file_id):
+    SharedUploadedFile = apps.get_model(
+        app_label='storage', model_name='SharedUploadedFile'
+    )
+
+    shared_uploaded_file = SharedUploadedFile.objects.get(
+        pk=shared_uploaded_file_id
+    )
+
+    try:
+        shared_uploaded_file.delete()
+    except OperationalError as exception:
+        logger.warning(
+            'Operational error attempting to delete shared upload file: '
+            '%s; %s. Retrying.', shared_uploaded_file, exception
+        )
+        raise self.retry(exc=exception)